Improved check for nameless function definitions
We use strlen() to check for the number of chars in argv[2] which is the
directory where the functions will be saved in. The *printf()-family of
functions will return the number of chars printed or -1 when they have failed.
This means that when we compare the return value of *printf()-family functions
to strlen() we need to make sure that the return value of *printf() is positive.
To also silence compiler warnings we cast the return value to size_t which is
the type returned by strlen().

Signed-off-by: Christian Brauner <christianvanbrauner@gmail.com>